A bit of refactoring of the Oceanus wrapper
